+++ ListDetail/app/src/main/java/com/microsoft/device/display/samples/listdetail/MainActivity.kt
@@ -5,18 +5,10 @@ import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
 import androidx.appcompat.app.AppCompatActivity
-import androidx.compose.foundation.text.BasicText
-import androidx.compose.material.Scaffold
-import androidx.compose.material.TopAppBar
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.setContent
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.sp
 import androidx.lifecycle.ViewModelProvider
 import androidx.window.WindowManager
-import com.microsoft.device.display.samples.companionpane.viewModels.AppStateViewModel
+import com.microsoft.device.display.samples.listdetail.models.AppStateViewModel
 import com.microsoft.device.display.samples.listdetail.ui.theme.ListDetailComposeSampleTheme
 import java.util.concurrent.Executor
 
@@ -36,25 +28,7 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContent {
             ListDetailComposeSampleTheme {
-                Scaffold(
-                    topBar = {
-                        TopAppBar(
-                            title = {
-                                BasicText(
-                                    text = stringResource(R.string.app_name),
-                                    style = TextStyle(
-                                        fontSize = 18.sp,
-                                        fontWeight = FontWeight.Bold,
-                                        color = Color.White
-                                    )
-                                )
-                            },
-                        )
-                    },
-                    bodyContent = {
-                        SetupUI(viewModel = appStateViewModel)
-                    }
-                )
+                SetupUI(appStateViewModel)
             }
         }
     }
